Move ADME and HTS to new files. Add HTS docs, tidy up datasets.rst (#430)
