+++ skfp/datasets/tdc/adme/__init__.py
@@ -1,25 +1,27 @@
-from .approved_pampa_ncats import load_approved_pampa_ncats
-from .b3db_classification import load_b3db_classification
-from .b3db_regression import load_b3db_regression
-from .bbb_martins import load_bbb_martins
-from .bioavailability_ma import load_bioavailability_ma
-from .caco2_wang import load_caco2_wang
-from .clearance_hepatocyte_az import load_clearance_hepatocyte_az
-from .clearance_microsome_az import load_clearance_microsome_az
-from .cyp1a2_veith import load_cyp1a2_veith
-from .cyp2c9_substrate_carbonmangels import load_cyp2c9_substrate_carbonmangels
-from .cyp2c9_veith import load_cyp2c9_veith
-from .cyp2c19_veith import load_cyp2c19_veith
-from .cyp2d6_substrate_carbonmangels import load_cyp2d6_substrate_carbonmangels
-from .cyp2d6_veith import load_cyp2d6_veith
-from .cyp3a4_substrate_carbonmangels import load_cyp3a4_substrate_carbonmangels
-from .cyp3a4_veith import load_cyp3a4_veith
-from .half_life_obach import load_half_life_obach
-from .hia_hou import load_hia_hou
-from .hlm import load_hlm
-from .pampa_ncats import load_pampa_ncats
-from .pgp_broccatelli import load_pgp_broccatelli
-from .ppbr_az import load_ppbr_az
-from .rlm import load_rlm
-from .solubility_aqsoldb import load_solubility_aqsoldb
-from .vdss_lombardo import load_vdss_lombardo
+from .adme import (
+    load_approved_pampa_ncats,
+    load_b3db_classification,
+    load_b3db_regression,
+    load_bbb_martins,
+    load_bioavailability_ma,
+    load_caco2_wang,
+    load_clearance_hepatocyte_az,
+    load_clearance_microsome_az,
+    load_cyp1a2_veith,
+    load_cyp2c9_substrate_carbonmangels,
+    load_cyp2c9_veith,
+    load_cyp2c19_veith,
+    load_cyp2d6_substrate_carbonmangels,
+    load_cyp2d6_veith,
+    load_cyp3a4_substrate_carbonmangels,
+    load_cyp3a4_veith,
+    load_half_life_obach,
+    load_hia_hou,
+    load_hlm,
+    load_pampa_ncats,
+    load_pgp_broccatelli,
+    load_ppbr_az,
+    load_rlm,
+    load_solubility_aqsoldb,
+    load_vdss_lombardo,
+)
