+++ skfp/datasets/tdc/adme/adme.py
@@ -1,0 +1,1183 @@
+import os
+from typing import Optional, Union
+
+import numpy as np
+import pandas as pd
+from sklearn.utils._param_validation import validate_params
+
+from skfp.datasets.utils import fetch_dataset, get_mol_strings_and_labels
+
+
+@validate_params(
+    {
+        "data_dir": [None, str, os.PathLike],
+        "as_frame": ["boolean"],
+        "verbose": ["boolean"],
+    },
+    prefer_skip_nested_validation=True,
+)
+def load_approved_pampa_ncats(
+    data_dir: Optional[Union[str, os.PathLike]] = None,
+    as_frame: bool = False,
+    verbose: bool = False,
+) -> Union[pd.DataFrame, tuple[list[str]], np.ndarray]:
+    """
+
+    Parameters
+    ----------
+    data_dir : {None, str, path-like}, default=None
+        Path to the root data directory. If ``None``, currently set scikit-learn directory
+        is used, by default `$HOME/scikit_learn_data`.
+
+    as_frame : bool, default=False
+        If True, returns the raw DataFrame with columns: "SMILES", "label". Otherwise,
+        returns SMILES as list of strings, and labels as a NumPy array (1D integer binary
+        vector).
+
+    verbose : bool, default=False
+        If True, progress bar will be shown for downloading or loading files.
+
+    Returns
+    -------
+    data : pd.DataFrame or tuple(list[str], np.ndarray)
+        Depending on the ``as_frame`` argument, one of:
+        - Pandas DataFrame with columns: "SMILES", "label"
+        - tuple of: list of strings (SMILES), NumPy array (labels)
+    -------
+
+    """
+    df = fetch_dataset(
+        data_dir,
+        dataset_name="TDC_approved_pampa_ncats",
+        filename="tdc_approved_pampa_ncats.csv",
+        verbose=verbose,
+    )
+    return df if as_frame else get_mol_strings_and_labels(df)
+
+
+@validate_params(
+    {
+        "data_dir": [None, str, os.PathLike],
+        "as_frame": ["boolean"],
+        "verbose": ["boolean"],
+    },
+    prefer_skip_nested_validation=True,
+)
+def load_b3db_classification(
+    data_dir: Optional[Union[str, os.PathLike]] = None,
+    as_frame: bool = False,
+    verbose: bool = False,
+) -> Union[pd.DataFrame, tuple[list[str]], np.ndarray]:
+    """
+
+    Parameters
+    ----------
+    data_dir : {None, str, path-like}, default=None
+        Path to the root data directory. If ``None``, currently set scikit-learn directory
+        is used, by default `$HOME/scikit_learn_data`.
+
+    as_frame : bool, default=False
+        If True, returns the raw DataFrame with columns: "SMILES", "label". Otherwise,
+        returns SMILES as list of strings, and labels as a NumPy array (1D integer binary
+        vector).
+
+    verbose : bool, default=False
+        If True, progress bar will be shown for downloading or loading files.
+
+    Returns
+    -------
+    data : pd.DataFrame or tuple(list[str], np.ndarray)
+        Depending on the ``as_frame`` argument, one of:
+        - Pandas DataFrame with columns: "SMILES", "label"
+        - tuple of: list of strings (SMILES), NumPy array (labels)
+    -------
+
+    """
+    df = fetch_dataset(
+        data_dir,
+        dataset_name="TDC_b3db_classification",
+        filename="tdc_b3db_classification.csv",
+        verbose=verbose,
+    )
+    return df if as_frame else get_mol_strings_and_labels(df)
+
+
+@validate_params(
+    {
+        "data_dir": [None, str, os.PathLike],
+        "as_frame": ["boolean"],
+        "verbose": ["boolean"],
+    },
+    prefer_skip_nested_validation=True,
+)
+def load_b3db_regression(
+    data_dir: Optional[Union[str, os.PathLike]] = None,
+    as_frame: bool = False,
+    verbose: bool = False,
+) -> Union[pd.DataFrame, tuple[list[str]], np.ndarray]:
+    """
+
+    Parameters
+    ----------
+    data_dir : {None, str, path-like}, default=None
+        Path to the root data directory. If ``None``, currently set scikit-learn directory
+        is used, by default `$HOME/scikit_learn_data`.
+
+    as_frame : bool, default=False
+        If True, returns the raw DataFrame with columns: "SMILES", "label". Otherwise,
+        returns SMILES as list of strings, and labels as a NumPy array (1D integer binary
+        vector).
+
+    verbose : bool, default=False
+        If True, progress bar will be shown for downloading or loading files.
+
+    Returns
+    -------
+    data : pd.DataFrame or tuple(list[str], np.ndarray)
+        Depending on the ``as_frame`` argument, one of:
+        - Pandas DataFrame with columns: "SMILES", "label"
+        - tuple of: list of strings (SMILES), NumPy array (labels)
+    -------
+
+    """
+    df = fetch_dataset(
+        data_dir,
+        dataset_name="TDC_b3db_regression",
+        filename="tdc_b3db_regression.csv",
+        verbose=verbose,
+    )
+    return df if as_frame else get_mol_strings_and_labels(df)
+
+
+@validate_params(
+    {
+        "data_dir": [None, str, os.PathLike],
+        "as_frame": ["boolean"],
+        "verbose": ["boolean"],
+    },
+    prefer_skip_nested_validation=True,
+)
+def load_bbb_martins(
+    data_dir: Optional[Union[str, os.PathLike]] = None,
+    as_frame: bool = False,
+    verbose: bool = False,
+) -> Union[pd.DataFrame, tuple[list[str]], np.ndarray]:
+    """
+
+    Parameters
+    ----------
+    data_dir : {None, str, path-like}, default=None
+        Path to the root data directory. If ``None``, currently set scikit-learn directory
+        is used, by default `$HOME/scikit_learn_data`.
+
+    as_frame : bool, default=False
+        If True, returns the raw DataFrame with columns: "SMILES", "label". Otherwise,
+        returns SMILES as list of strings, and labels as a NumPy array (1D integer binary
+        vector).
+
+    verbose : bool, default=False
+        If True, progress bar will be shown for downloading or loading files.
+
+    Returns
+    -------
+    data : pd.DataFrame or tuple(list[str], np.ndarray)
+        Depending on the ``as_frame`` argument, one of:
+        - Pandas DataFrame with columns: "SMILES", "label"
+        - tuple of: list of strings (SMILES), NumPy array (labels)
+    -------
+
+    """
+    df = fetch_dataset(
+        data_dir,
+        dataset_name="TDC_bbb_martins",
+        filename="tdc_bbb_martins.csv",
+        verbose=verbose,
+    )
+    return df if as_frame else get_mol_strings_and_labels(df)
+
+
+@validate_params(
+    {
+        "data_dir": [None, str, os.PathLike],
+        "as_frame": ["boolean"],
+        "verbose": ["boolean"],
+    },
+    prefer_skip_nested_validation=True,
+)
+def load_bioavailability_ma(
+    data_dir: Optional[Union[str, os.PathLike]] = None,
+    as_frame: bool = False,
+    verbose: bool = False,
+) -> Union[pd.DataFrame, tuple[list[str]], np.ndarray]:
+    """
+
+    Parameters
+    ----------
+    data_dir : {None, str, path-like}, default=None
+        Path to the root data directory. If ``None``, currently set scikit-learn directory
+        is used, by default `$HOME/scikit_learn_data`.
+
+    as_frame : bool, default=False
+        If True, returns the raw DataFrame with columns: "SMILES", "label". Otherwise,
+        returns SMILES as list of strings, and labels as a NumPy array (1D integer binary
+        vector).
+
+    verbose : bool, default=False
+        If True, progress bar will be shown for downloading or loading files.
+
+    Returns
+    -------
+    data : pd.DataFrame or tuple(list[str], np.ndarray)
+        Depending on the ``as_frame`` argument, one of:
+        - Pandas DataFrame with columns: "SMILES", "label"
+        - tuple of: list of strings (SMILES), NumPy array (labels)
+    -------
+
+    """
+    df = fetch_dataset(
+        data_dir,
+        dataset_name="TDC_bioavailability_ma",
+        filename="tdc_bioavailability_ma.csv",
+        verbose=verbose,
+    )
+    return df if as_frame else get_mol_strings_and_labels(df)
+
+
+@validate_params(
+    {
+        "data_dir": [None, str, os.PathLike],
+        "as_frame": ["boolean"],
+        "verbose": ["boolean"],
+    },
+    prefer_skip_nested_validation=True,
+)
+def load_caco2_wang(
+    data_dir: Optional[Union[str, os.PathLike]] = None,
+    as_frame: bool = False,
+    verbose: bool = False,
+) -> Union[pd.DataFrame, tuple[list[str]], np.ndarray]:
+    """
+
+    Parameters
+    ----------
+    data_dir : {None, str, path-like}, default=None
+        Path to the root data directory. If ``None``, currently set scikit-learn directory
+        is used, by default `$HOME/scikit_learn_data`.
+
+    as_frame : bool, default=False
+        If True, returns the raw DataFrame with columns: "SMILES", "label". Otherwise,
+        returns SMILES as list of strings, and labels as a NumPy array (1D integer binary
+        vector).
+
+    verbose : bool, default=False
+        If True, progress bar will be shown for downloading or loading files.
+
+    Returns
+    -------
+    data : pd.DataFrame or tuple(list[str], np.ndarray)
+        Depending on the ``as_frame`` argument, one of:
+        - Pandas DataFrame with columns: "SMILES", "label"
+        - tuple of: list of strings (SMILES), NumPy array (labels)
+    -------
+
+    """
+    df = fetch_dataset(
+        data_dir,
+        dataset_name="TDC_caco2_wang",
+        filename="tdc_caco2_wang.csv",
+        verbose=verbose,
+    )
+    return df if as_frame else get_mol_strings_and_labels(df)
+
+
+@validate_params(
+    {
+        "data_dir": [None, str, os.PathLike],
+        "as_frame": ["boolean"],
+        "verbose": ["boolean"],
+    },
+    prefer_skip_nested_validation=True,
+)
+def load_clearance_hepatocyte_az(
+    data_dir: Optional[Union[str, os.PathLike]] = None,
+    as_frame: bool = False,
+    verbose: bool = False,
+) -> Union[pd.DataFrame, tuple[list[str]], np.ndarray]:
+    """
+
+    Parameters
+    ----------
+    data_dir : {None, str, path-like}, default=None
+        Path to the root data directory. If ``None``, currently set scikit-learn directory
+        is used, by default `$HOME/scikit_learn_data`.
+
+    as_frame : bool, default=False
+        If True, returns the raw DataFrame with columns: "SMILES", "label". Otherwise,
+        returns SMILES as list of strings, and labels as a NumPy array (1D integer binary
+        vector).
+
+    verbose : bool, default=False
+        If True, progress bar will be shown for downloading or loading files.
+
+    Returns
+    -------
+    data : pd.DataFrame or tuple(list[str], np.ndarray)
+        Depending on the ``as_frame`` argument, one of:
+        - Pandas DataFrame with columns: "SMILES", "label"
+        - tuple of: list of strings (SMILES), NumPy array (labels)
+    -------
+
+    """
+    df = fetch_dataset(
+        data_dir,
+        dataset_name="TDC_clearance_hepatocyte_az",
+        filename="tdc_clearance_hepatocyte_az.csv",
+        verbose=verbose,
+    )
+    return df if as_frame else get_mol_strings_and_labels(df)
+
+
+@validate_params(
+    {
+        "data_dir": [None, str, os.PathLike],
+        "as_frame": ["boolean"],
+        "verbose": ["boolean"],
+    },
+    prefer_skip_nested_validation=True,
+)
+def load_clearance_microsome_az(
+    data_dir: Optional[Union[str, os.PathLike]] = None,
+    as_frame: bool = False,
+    verbose: bool = False,
+) -> Union[pd.DataFrame, tuple[list[str]], np.ndarray]:
+    """
+
+    Parameters
+    ----------
+    data_dir : {None, str, path-like}, default=None
+        Path to the root data directory. If ``None``, currently set scikit-learn directory
+        is used, by default `$HOME/scikit_learn_data`.
+
+    as_frame : bool, default=False
+        If True, returns the raw DataFrame with columns: "SMILES", "label". Otherwise,
+        returns SMILES as list of strings, and labels as a NumPy array (1D integer binary
+        vector).
+
+    verbose : bool, default=False
+        If True, progress bar will be shown for downloading or loading files.
+
+    Returns
+    -------
+    data : pd.DataFrame or tuple(list[str], np.ndarray)
+        Depending on the ``as_frame`` argument, one of:
+        - Pandas DataFrame with columns: "SMILES", "label"
+        - tuple of: list of strings (SMILES), NumPy array (labels)
+    -------
+
+    """
+    df = fetch_dataset(
+        data_dir,
+        dataset_name="TDC_clearance_microsome_az",
+        filename="tdc_clearance_microsome_az.csv",
+        verbose=verbose,
+    )
+    return df if as_frame else get_mol_strings_and_labels(df)
+
+
+@validate_params(
+    {
+        "data_dir": [None, str, os.PathLike],
+        "as_frame": ["boolean"],
+        "verbose": ["boolean"],
+    },
+    prefer_skip_nested_validation=True,
+)
+def load_cyp1a2_veith(
+    data_dir: Optional[Union[str, os.PathLike]] = None,
+    as_frame: bool = False,
+    verbose: bool = False,
+) -> Union[pd.DataFrame, tuple[list[str]], np.ndarray]:
+    """
+
+    Parameters
+    ----------
+    data_dir : {None, str, path-like}, default=None
+        Path to the root data directory. If ``None``, currently set scikit-learn directory
+        is used, by default `$HOME/scikit_learn_data`.
+
+    as_frame : bool, default=False
+        If True, returns the raw DataFrame with columns: "SMILES", "label". Otherwise,
+        returns SMILES as list of strings, and labels as a NumPy array (1D integer binary
+        vector).
+
+    verbose : bool, default=False
+        If True, progress bar will be shown for downloading or loading files.
+
+    Returns
+    -------
+    data : pd.DataFrame or tuple(list[str], np.ndarray)
+        Depending on the ``as_frame`` argument, one of:
+        - Pandas DataFrame with columns: "SMILES", "label"
+        - tuple of: list of strings (SMILES), NumPy array (labels)
+    -------
+
+    """
+    df = fetch_dataset(
+        data_dir,
+        dataset_name="TDC_cyp1a2_veith",
+        filename="tdc_cyp1a2_veith.csv",
+        verbose=verbose,
+    )
+    return df if as_frame else get_mol_strings_and_labels(df)
+
+
+@validate_params(
+    {
+        "data_dir": [None, str, os.PathLike],
+        "as_frame": ["boolean"],
+        "verbose": ["boolean"],
+    },
+    prefer_skip_nested_validation=True,
+)
+def load_cyp2c9_substrate_carbonmangels(
+    data_dir: Optional[Union[str, os.PathLike]] = None,
+    as_frame: bool = False,
+    verbose: bool = False,
+) -> Union[pd.DataFrame, tuple[list[str]], np.ndarray]:
+    """
+
+    Parameters
+    ----------
+    data_dir : {None, str, path-like}, default=None
+        Path to the root data directory. If ``None``, currently set scikit-learn directory
+        is used, by default `$HOME/scikit_learn_data`.
+
+    as_frame : bool, default=False
+        If True, returns the raw DataFrame with columns: "SMILES", "label". Otherwise,
+        returns SMILES as list of strings, and labels as a NumPy array (1D integer binary
+        vector).
+
+    verbose : bool, default=False
+        If True, progress bar will be shown for downloading or loading files.
+
+    Returns
+    -------
+    data : pd.DataFrame or tuple(list[str], np.ndarray)
+        Depending on the ``as_frame`` argument, one of:
+        - Pandas DataFrame with columns: "SMILES", "label"
+        - tuple of: list of strings (SMILES), NumPy array (labels)
+    -------
+
+    """
+    df = fetch_dataset(
+        data_dir,
+        dataset_name="TDC_cyp2c9_substrate_carbonmangels",
+        filename="tdc_cyp2c9_substrate_carbonmangels.csv",
+        verbose=verbose,
+    )
+    return df if as_frame else get_mol_strings_and_labels(df)
+
+
+@validate_params(
+    {
+        "data_dir": [None, str, os.PathLike],
+        "as_frame": ["boolean"],
+        "verbose": ["boolean"],
+    },
+    prefer_skip_nested_validation=True,
+)
+def load_cyp2c9_veith(
+    data_dir: Optional[Union[str, os.PathLike]] = None,
+    as_frame: bool = False,
+    verbose: bool = False,
+) -> Union[pd.DataFrame, tuple[list[str]], np.ndarray]:
+    """
+
+    Parameters
+    ----------
+    data_dir : {None, str, path-like}, default=None
+        Path to the root data directory. If ``None``, currently set scikit-learn directory
+        is used, by default `$HOME/scikit_learn_data`.
+
+    as_frame : bool, default=False
+        If True, returns the raw DataFrame with columns: "SMILES", "label". Otherwise,
+        returns SMILES as list of strings, and labels as a NumPy array (1D integer binary
+        vector).
+
+    verbose : bool, default=False
+        If True, progress bar will be shown for downloading or loading files.
+
+    Returns
+    -------
+    data : pd.DataFrame or tuple(list[str], np.ndarray)
+        Depending on the ``as_frame`` argument, one of:
+        - Pandas DataFrame with columns: "SMILES", "label"
+        - tuple of: list of strings (SMILES), NumPy array (labels)
+    -------
+
+    """
+    df = fetch_dataset(
+        data_dir,
+        dataset_name="TDC_cyp2c9_veith",
+        filename="tdc_cyp2c9_veith.csv",
+        verbose=verbose,
+    )
+    return df if as_frame else get_mol_strings_and_labels(df)
+
+
+@validate_params(
+    {
+        "data_dir": [None, str, os.PathLike],
+        "as_frame": ["boolean"],
+        "verbose": ["boolean"],
+    },
+    prefer_skip_nested_validation=True,
+)
+def load_cyp2c19_veith(
+    data_dir: Optional[Union[str, os.PathLike]] = None,
+    as_frame: bool = False,
+    verbose: bool = False,
+) -> Union[pd.DataFrame, tuple[list[str]], np.ndarray]:
+    """
+
+    Parameters
+    ----------
+    data_dir : {None, str, path-like}, default=None
+        Path to the root data directory. If ``None``, currently set scikit-learn directory
+        is used, by default `$HOME/scikit_learn_data`.
+
+    as_frame : bool, default=False
+        If True, returns the raw DataFrame with columns: "SMILES", "label". Otherwise,
+        returns SMILES as list of strings, and labels as a NumPy array (1D integer binary
+        vector).
+
+    verbose : bool, default=False
+        If True, progress bar will be shown for downloading or loading files.
+
+    Returns
+    -------
+    data : pd.DataFrame or tuple(list[str], np.ndarray)
+        Depending on the ``as_frame`` argument, one of:
+        - Pandas DataFrame with columns: "SMILES", "label"
+        - tuple of: list of strings (SMILES), NumPy array (labels)
+    -------
+
+    """
+    df = fetch_dataset(
+        data_dir,
+        dataset_name="TDC_cyp2c19_veith",
+        filename="tdc_cyp2c19_veith.csv",
+        verbose=verbose,
+    )
+    return df if as_frame else get_mol_strings_and_labels(df)
+
+
+@validate_params(
+    {
+        "data_dir": [None, str, os.PathLike],
+        "as_frame": ["boolean"],
+        "verbose": ["boolean"],
+    },
+    prefer_skip_nested_validation=True,
+)
+def load_cyp2d6_substrate_carbonmangels(
+    data_dir: Optional[Union[str, os.PathLike]] = None,
+    as_frame: bool = False,
+    verbose: bool = False,
+) -> Union[pd.DataFrame, tuple[list[str]], np.ndarray]:
+    """
+
+    Parameters
+    ----------
+    data_dir : {None, str, path-like}, default=None
+        Path to the root data directory. If ``None``, currently set scikit-learn directory
+        is used, by default `$HOME/scikit_learn_data`.
+
+    as_frame : bool, default=False
+        If True, returns the raw DataFrame with columns: "SMILES", "label". Otherwise,
+        returns SMILES as list of strings, and labels as a NumPy array (1D integer binary
+        vector).
+
+    verbose : bool, default=False
+        If True, progress bar will be shown for downloading or loading files.
+
+    Returns
+    -------
+    data : pd.DataFrame or tuple(list[str], np.ndarray)
+        Depending on the ``as_frame`` argument, one of:
+        - Pandas DataFrame with columns: "SMILES", "label"
+        - tuple of: list of strings (SMILES), NumPy array (labels)
+    -------
+
+    """
+    df = fetch_dataset(
+        data_dir,
+        dataset_name="TDC_cyp2d6_substrate_carbonmangels",
+        filename="tdc_cyp2d6_substrate_carbonmangels.csv",
+        verbose=verbose,
+    )
+    return df if as_frame else get_mol_strings_and_labels(df)
+
+
+@validate_params(
+    {
+        "data_dir": [None, str, os.PathLike],
+        "as_frame": ["boolean"],
+        "verbose": ["boolean"],
+    },
+    prefer_skip_nested_validation=True,
+)
+def load_cyp2d6_veith(
+    data_dir: Optional[Union[str, os.PathLike]] = None,
+    as_frame: bool = False,
+    verbose: bool = False,
+) -> Union[pd.DataFrame, tuple[list[str]], np.ndarray]:
+    """
+
+    Parameters
+    ----------
+    data_dir : {None, str, path-like}, default=None
+        Path to the root data directory. If ``None``, currently set scikit-learn directory
+        is used, by default `$HOME/scikit_learn_data`.
+
+    as_frame : bool, default=False
+        If True, returns the raw DataFrame with columns: "SMILES", "label". Otherwise,
+        returns SMILES as list of strings, and labels as a NumPy array (1D integer binary
+        vector).
+
+    verbose : bool, default=False
+        If True, progress bar will be shown for downloading or loading files.
+
+    Returns
+    -------
+    data : pd.DataFrame or tuple(list[str], np.ndarray)
+        Depending on the ``as_frame`` argument, one of:
+        - Pandas DataFrame with columns: "SMILES", "label"
+        - tuple of: list of strings (SMILES), NumPy array (labels)
+    -------
+
+    """
+    df = fetch_dataset(
+        data_dir,
+        dataset_name="TDC_cyp2d6_veith",
+        filename="tdc_cyp2d6_veith.csv",
+        verbose=verbose,
+    )
+    return df if as_frame else get_mol_strings_and_labels(df)
+
+
+@validate_params(
+    {
+        "data_dir": [None, str, os.PathLike],
+        "as_frame": ["boolean"],
+        "verbose": ["boolean"],
+    },
+    prefer_skip_nested_validation=True,
+)
+def load_cyp3a4_substrate_carbonmangels(
+    data_dir: Optional[Union[str, os.PathLike]] = None,
+    as_frame: bool = False,
+    verbose: bool = False,
+) -> Union[pd.DataFrame, tuple[list[str]], np.ndarray]:
+    """
+
+    Parameters
+    ----------
+    data_dir : {None, str, path-like}, default=None
+        Path to the root data directory. If ``None``, currently set scikit-learn directory
+        is used, by default `$HOME/scikit_learn_data`.
+
+    as_frame : bool, default=False
+        If True, returns the raw DataFrame with columns: "SMILES", "label". Otherwise,
+        returns SMILES as list of strings, and labels as a NumPy array (1D integer binary
+        vector).
+
+    verbose : bool, default=False
+        If True, progress bar will be shown for downloading or loading files.
+
+    Returns
+    -------
+    data : pd.DataFrame or tuple(list[str], np.ndarray)
+        Depending on the ``as_frame`` argument, one of:
+        - Pandas DataFrame with columns: "SMILES", "label"
+        - tuple of: list of strings (SMILES), NumPy array (labels)
+    -------
+
+    """
+    df = fetch_dataset(
+        data_dir,
+        dataset_name="TDC_cyp3a4_substrate_carbonmangels",
+        filename="tdc_cyp3a4_substrate_carbonmangels.csv",
+        verbose=verbose,
+    )
+    return df if as_frame else get_mol_strings_and_labels(df)
+
+
+@validate_params(
+    {
+        "data_dir": [None, str, os.PathLike],
+        "as_frame": ["boolean"],
+        "verbose": ["boolean"],
+    },
+    prefer_skip_nested_validation=True,
+)
+def load_cyp3a4_veith(
+    data_dir: Optional[Union[str, os.PathLike]] = None,
+    as_frame: bool = False,
+    verbose: bool = False,
+) -> Union[pd.DataFrame, tuple[list[str]], np.ndarray]:
+    """
+
+    Parameters
+    ----------
+    data_dir : {None, str, path-like}, default=None
+        Path to the root data directory. If ``None``, currently set scikit-learn directory
+        is used, by default `$HOME/scikit_learn_data`.
+
+    as_frame : bool, default=False
+        If True, returns the raw DataFrame with columns: "SMILES", "label". Otherwise,
+        returns SMILES as list of strings, and labels as a NumPy array (1D integer binary
+        vector).
+
+    verbose : bool, default=False
+        If True, progress bar will be shown for downloading or loading files.
+
+    Returns
+    -------
+    data : pd.DataFrame or tuple(list[str], np.ndarray)
+        Depending on the ``as_frame`` argument, one of:
+        - Pandas DataFrame with columns: "SMILES", "label"
+        - tuple of: list of strings (SMILES), NumPy array (labels)
+    -------
+
+    """
+    df = fetch_dataset(
+        data_dir,
+        dataset_name="TDC_cyp3a4_veith",
+        filename="tdc_cyp3a4_veith.csv",
+        verbose=verbose,
+    )
+    return df if as_frame else get_mol_strings_and_labels(df)
+
+
+@validate_params(
+    {
+        "data_dir": [None, str, os.PathLike],
+        "as_frame": ["boolean"],
+        "verbose": ["boolean"],
+    },
+    prefer_skip_nested_validation=True,
+)
+def load_half_life_obach(
+    data_dir: Optional[Union[str, os.PathLike]] = None,
+    as_frame: bool = False,
+    verbose: bool = False,
+) -> Union[pd.DataFrame, tuple[list[str]], np.ndarray]:
+    """
+
+    Parameters
+    ----------
+    data_dir : {None, str, path-like}, default=None
+        Path to the root data directory. If ``None``, currently set scikit-learn directory
+        is used, by default `$HOME/scikit_learn_data`.
+
+    as_frame : bool, default=False
+        If True, returns the raw DataFrame with columns: "SMILES", "label". Otherwise,
+        returns SMILES as list of strings, and labels as a NumPy array (1D integer binary
+        vector).
+
+    verbose : bool, default=False
+        If True, progress bar will be shown for downloading or loading files.
+
+    Returns
+    -------
+    data : pd.DataFrame or tuple(list[str], np.ndarray)
+        Depending on the ``as_frame`` argument, one of:
+        - Pandas DataFrame with columns: "SMILES", "label"
+        - tuple of: list of strings (SMILES), NumPy array (labels)
+    -------
+
+    """
+    df = fetch_dataset(
+        data_dir,
+        dataset_name="TDC_half_life_obach",
+        filename="tdc_half_life_obach.csv",
+        verbose=verbose,
+    )
+    return df if as_frame else get_mol_strings_and_labels(df)
+
+
+@validate_params(
+    {
+        "data_dir": [None, str, os.PathLike],
+        "as_frame": ["boolean"],
+        "verbose": ["boolean"],
+    },
+    prefer_skip_nested_validation=True,
+)
+def load_hia_hou(
+    data_dir: Optional[Union[str, os.PathLike]] = None,
+    as_frame: bool = False,
+    verbose: bool = False,
+) -> Union[pd.DataFrame, tuple[list[str]], np.ndarray]:
+    """
+
+    Parameters
+    ----------
+    data_dir : {None, str, path-like}, default=None
+        Path to the root data directory. If ``None``, currently set scikit-learn directory
+        is used, by default `$HOME/scikit_learn_data`.
+
+    as_frame : bool, default=False
+        If True, returns the raw DataFrame with columns: "SMILES", "label". Otherwise,
+        returns SMILES as list of strings, and labels as a NumPy array (1D integer binary
+        vector).
+
+    verbose : bool, default=False
+        If True, progress bar will be shown for downloading or loading files.
+
+    Returns
+    -------
+    data : pd.DataFrame or tuple(list[str], np.ndarray)
+        Depending on the ``as_frame`` argument, one of:
+        - Pandas DataFrame with columns: "SMILES", "label"
+        - tuple of: list of strings (SMILES), NumPy array (labels)
+    -------
+
+    """
+    df = fetch_dataset(
+        data_dir,
+        dataset_name="TDC_hia_hou",
+        filename="tdc_hia_hou.csv",
+        verbose=verbose,
+    )
+    return df if as_frame else get_mol_strings_and_labels(df)
+
+
+@validate_params(
+    {
+        "data_dir": [None, str, os.PathLike],
+        "as_frame": ["boolean"],
+        "verbose": ["boolean"],
+    },
+    prefer_skip_nested_validation=True,
+)
+def load_hlm(
+    data_dir: Optional[Union[str, os.PathLike]] = None,
+    as_frame: bool = False,
+    verbose: bool = False,
+) -> Union[pd.DataFrame, tuple[list[str]], np.ndarray]:
+    """
+
+    Parameters
+    ----------
+    data_dir : {None, str, path-like}, default=None
+        Path to the root data directory. If ``None``, currently set scikit-learn directory
+        is used, by default `$HOME/scikit_learn_data`.
+
+    as_frame : bool, default=False
+        If True, returns the raw DataFrame with columns: "SMILES", "label". Otherwise,
+        returns SMILES as list of strings, and labels as a NumPy array (1D integer binary
+        vector).
+
+    verbose : bool, default=False
+        If True, progress bar will be shown for downloading or loading files.
+
+    Returns
+    -------
+    data : pd.DataFrame or tuple(list[str], np.ndarray)
+        Depending on the ``as_frame`` argument, one of:
+        - Pandas DataFrame with columns: "SMILES", "label"
+        - tuple of: list of strings (SMILES), NumPy array (labels)
+    -------
+
+    """
+    df = fetch_dataset(
+        data_dir,
+        dataset_name="TDC_hlm",
+        filename="tdc_hlm.csv",
+        verbose=verbose,
+    )
+    return df if as_frame else get_mol_strings_and_labels(df)
+
+
+@validate_params(
+    {
+        "data_dir": [None, str, os.PathLike],
+        "as_frame": ["boolean"],
+        "verbose": ["boolean"],
+    },
+    prefer_skip_nested_validation=True,
+)
+def load_pampa_ncats(
+    data_dir: Optional[Union[str, os.PathLike]] = None,
+    as_frame: bool = False,
+    verbose: bool = False,
+) -> Union[pd.DataFrame, tuple[list[str]], np.ndarray]:
+    """
+
+    Parameters
+    ----------
+    data_dir : {None, str, path-like}, default=None
+        Path to the root data directory. If ``None``, currently set scikit-learn directory
+        is used, by default `$HOME/scikit_learn_data`.
+
+    as_frame : bool, default=False
+        If True, returns the raw DataFrame with columns: "SMILES", "label". Otherwise,
+        returns SMILES as list of strings, and labels as a NumPy array (1D integer binary
+        vector).
+
+    verbose : bool, default=False
+        If True, progress bar will be shown for downloading or loading files.
+
+    Returns
+    -------
+    data : pd.DataFrame or tuple(list[str], np.ndarray)
+        Depending on the ``as_frame`` argument, one of:
+        - Pandas DataFrame with columns: "SMILES", "label"
+        - tuple of: list of strings (SMILES), NumPy array (labels)
+    -------
+
+    """
+    df = fetch_dataset(
+        data_dir,
+        dataset_name="TDC_pampa_ncats",
+        filename="tdc_pampa_ncats.csv",
+        verbose=verbose,
+    )
+    return df if as_frame else get_mol_strings_and_labels(df)
+
+
+@validate_params(
+    {
+        "data_dir": [None, str, os.PathLike],
+        "as_frame": ["boolean"],
+        "verbose": ["boolean"],
+    },
+    prefer_skip_nested_validation=True,
+)
+def load_pgp_broccatelli(
+    data_dir: Optional[Union[str, os.PathLike]] = None,
+    as_frame: bool = False,
+    verbose: bool = False,
+) -> Union[pd.DataFrame, tuple[list[str]], np.ndarray]:
+    """
+
+    Parameters
+    ----------
+    data_dir : {None, str, path-like}, default=None
+        Path to the root data directory. If ``None``, currently set scikit-learn directory
+        is used, by default `$HOME/scikit_learn_data`.
+
+    as_frame : bool, default=False
+        If True, returns the raw DataFrame with columns: "SMILES", "label". Otherwise,
+        returns SMILES as list of strings, and labels as a NumPy array (1D integer binary
+        vector).
+
+    verbose : bool, default=False
+        If True, progress bar will be shown for downloading or loading files.
+
+    Returns
+    -------
+    data : pd.DataFrame or tuple(list[str], np.ndarray)
+        Depending on the ``as_frame`` argument, one of:
+        - Pandas DataFrame with columns: "SMILES", "label"
+        - tuple of: list of strings (SMILES), NumPy array (labels)
+    -------
+
+    """
+    df = fetch_dataset(
+        data_dir,
+        dataset_name="TDC_pgp_broccatelli",
+        filename="tdc_pgp_broccatelli.csv",
+        verbose=verbose,
+    )
+    return df if as_frame else get_mol_strings_and_labels(df)
+
+
+@validate_params(
+    {
+        "data_dir": [None, str, os.PathLike],
+        "as_frame": ["boolean"],
+        "verbose": ["boolean"],
+    },
+    prefer_skip_nested_validation=True,
+)
+def load_ppbr_az(
+    data_dir: Optional[Union[str, os.PathLike]] = None,
+    as_frame: bool = False,
+    verbose: bool = False,
+) -> Union[pd.DataFrame, tuple[list[str]], np.ndarray]:
+    """
+
+    Parameters
+    ----------
+    data_dir : {None, str, path-like}, default=None
+        Path to the root data directory. If ``None``, currently set scikit-learn directory
+        is used, by default `$HOME/scikit_learn_data`.
+
+    as_frame : bool, default=False
+        If True, returns the raw DataFrame with columns: "SMILES", "label". Otherwise,
+        returns SMILES as list of strings, and labels as a NumPy array (1D integer binary
+        vector).
+
+    verbose : bool, default=False
+        If True, progress bar will be shown for downloading or loading files.
+
+    Returns
+    -------
+    data : pd.DataFrame or tuple(list[str], np.ndarray)
+        Depending on the ``as_frame`` argument, one of:
+        - Pandas DataFrame with columns: "SMILES", "label"
+        - tuple of: list of strings (SMILES), NumPy array (labels)
+    -------
+
+    """
+    df = fetch_dataset(
+        data_dir,
+        dataset_name="TDC_ppbr_az",
+        filename="tdc_ppbr_az.csv",
+        verbose=verbose,
+    )
+    return df if as_frame else get_mol_strings_and_labels(df)
+
+
+@validate_params(
+    {
+        "data_dir": [None, str, os.PathLike],
+        "as_frame": ["boolean"],
+        "verbose": ["boolean"],
+    },
+    prefer_skip_nested_validation=True,
+)
+def load_rlm(
+    data_dir: Optional[Union[str, os.PathLike]] = None,
+    as_frame: bool = False,
+    verbose: bool = False,
+) -> Union[pd.DataFrame, tuple[list[str]], np.ndarray]:
+    """
+
+    Parameters
+    ----------
+    data_dir : {None, str, path-like}, default=None
+        Path to the root data directory. If ``None``, currently set scikit-learn directory
+        is used, by default `$HOME/scikit_learn_data`.
+
+    as_frame : bool, default=False
+        If True, returns the raw DataFrame with columns: "SMILES", "label". Otherwise,
+        returns SMILES as list of strings, and labels as a NumPy array (1D integer binary
+        vector).
+
+    verbose : bool, default=False
+        If True, progress bar will be shown for downloading or loading files.
+
+    Returns
+    -------
+    data : pd.DataFrame or tuple(list[str], np.ndarray)
+        Depending on the ``as_frame`` argument, one of:
+        - Pandas DataFrame with columns: "SMILES", "label"
+        - tuple of: list of strings (SMILES), NumPy array (labels)
+    -------
+
+    """
+    df = fetch_dataset(
+        data_dir,
+        dataset_name="TDC_rlm",
+        filename="tdc_rlm.csv",
+        verbose=verbose,
+    )
+    return df if as_frame else get_mol_strings_and_labels(df)
+
+
+@validate_params(
+    {
+        "data_dir": [None, str, os.PathLike],
+        "as_frame": ["boolean"],
+        "verbose": ["boolean"],
+    },
+    prefer_skip_nested_validation=True,
+)
+def load_solubility_aqsoldb(
+    data_dir: Optional[Union[str, os.PathLike]] = None,
+    as_frame: bool = False,
+    verbose: bool = False,
+) -> Union[pd.DataFrame, tuple[list[str]], np.ndarray]:
+    """
+
+    Parameters
+    ----------
+    data_dir : {None, str, path-like}, default=None
+        Path to the root data directory. If ``None``, currently set scikit-learn directory
+        is used, by default `$HOME/scikit_learn_data`.
+
+    as_frame : bool, default=False
+        If True, returns the raw DataFrame with columns: "SMILES", "label". Otherwise,
+        returns SMILES as list of strings, and labels as a NumPy array (1D integer binary
+        vector).
+
+    verbose : bool, default=False
+        If True, progress bar will be shown for downloading or loading files.
+
+    Returns
+    -------
+    data : pd.DataFrame or tuple(list[str], np.ndarray)
+        Depending on the ``as_frame`` argument, one of:
+        - Pandas DataFrame with columns: "SMILES", "label"
+        - tuple of: list of strings (SMILES), NumPy array (labels)
+    -------
+
+    """
+    df = fetch_dataset(
+        data_dir,
+        dataset_name="TDC_solubility_aqsoldb",
+        filename="tdc_solubility_aqsoldb.csv",
+        verbose=verbose,
+    )
+    return df if as_frame else get_mol_strings_and_labels(df)
+
+
+@validate_params(
+    {
+        "data_dir": [None, str, os.PathLike],
+        "as_frame": ["boolean"],
+        "verbose": ["boolean"],
+    },
+    prefer_skip_nested_validation=True,
+)
+def load_vdss_lombardo(
+    data_dir: Optional[Union[str, os.PathLike]] = None,
+    as_frame: bool = False,
+    verbose: bool = False,
+) -> Union[pd.DataFrame, tuple[list[str]], np.ndarray]:
+    """
+
+    Parameters
+    ----------
+    data_dir : {None, str, path-like}, default=None
+        Path to the root data directory. If ``None``, currently set scikit-learn directory
+        is used, by default `$HOME/scikit_learn_data`.
+
+    as_frame : bool, default=False
+        If True, returns the raw DataFrame with columns: "SMILES", "label". Otherwise,
+        returns SMILES as list of strings, and labels as a NumPy array (1D integer binary
+        vector).
+
+    verbose : bool, default=False
+        If True, progress bar will be shown for downloading or loading files.
+
+    Returns
+    -------
+    data : pd.DataFrame or tuple(list[str], np.ndarray)
+        Depending on the ``as_frame`` argument, one of:
+        - Pandas DataFrame with columns: "SMILES", "label"
+        - tuple of: list of strings (SMILES), NumPy array (labels)
+    -------
+
+    """
+    df = fetch_dataset(
+        data_dir,
+        dataset_name="TDC_vdss_lombardo",
+        filename="tdc_vdss_lombardo.csv",
+        verbose=verbose,
+    )
+    return df if as_frame else get_mol_strings_and_labels(df)
